remove check for dummy texture IID and rename DLL
The check for the dummy IID did not have an effect since the loop
eventually returned the same result, even without the dummy texture on a
lot. As this also simplifies usage of the DLL, the dummy IID was removed
altogether, so that existing lots do not need to be updated to make use
of the fix.

This does mean that the game sometimes renders terrain below base
textures unnecessarily, but that's already the case before this fix (at
least, whenever you leave the Zoning data view).